Split out dependencies into separate modules
